Extract placeholder-ledger bookkeeping into a mixin (WS5 Seam 1)
Move the 35 ignored-active-replay placeholder / dependent-reply bookkeeping
methods out of LCMEngine into a PlaceholderLedgerMixin in the new
placeholder_ledger.py, and mix it into LCMEngine
(`class LCMEngine(ContextEngine, PlaceholderLedgerMixin)`).

The methods are moved verbatim (byte-identical bodies) and keep running bound to
the engine instance: `self` is the LCMEngine, so they read/write the engine's
shared runtime state (`_store`, `_session_id`, the two
`_generated_ignored_active_replay_placeholder_*` sets, the per-turn
`_current_compress_store_ids_by_message_id` cache) and call engine helpers
(`_get_store_id_map_for_messages`, `_matches_ignore_message_patterns`,
`_stable_tool_calls_identity`, `_message_replay_identity`,
`_copy_active_replay_messages_preserving_generated_ids`) through normal attribute
lookup via the MRO. No call sites and no tests change; the ~130 white-box tests
in tests/test_lcm_engine.py exercise the methods unchanged.

engine.py shrinks by ~657 lines. The two shared placeholder-id sets and the
per-turn store-id cache stay owned by LCMEngine (they are also touched by
sanitize/reconcile methods that remain on the engine), so the mixin is a pure
relocation, not a state-ownership change.

Behaviour-preserving refactor (WS5 engine decomposition). Rollback: git revert
this commit re-inlines the methods; no data or state migration involved.

Bases order: the mixin precedes ContextEngine so the LCM decomposition mixins
resolve consistently across seams (required for CompactionMixin, which overrides
the compress/should_compress protocol methods; harmless for private-only mixins).
